feat: Add resiliency support for WS
